Struct rsix::path::DecInt

pub struct DecInt { /* fields omitted */ }

Format an integer into a decimal Path component, without constructing a temporary PathBuf or String.

This is used for opening paths such as /proc/self/fd/<fd> on Linux.

Example

use rsix::path::DecInt;

assert_eq!(
    format!("hello {}", DecInt::new(9876).display()),
    "hello 9876"
);

Implementations

impl DecInt

pub fn new<Int: Integer>(i: Int) -> Self

Construct a new path component from an integer.

pub fn from_fd<Fd: AsFd>(fd: &Fd) -> Self

Construct a new path component from a file descriptor.

pub fn as_bytes(&self) -> &[u8]

Return the raw byte buffer.

pub fn as_c_str(&self) -> &CStr

Return the raw byte buffer.